# src/main/resources/static/petanque.openapi.yaml
openapi: 3.0.1
info:
  title: Petanque
  version: 0.0.0
  description: Keep score and ranking for petanque players
  contact:
    name: Olivier Francome
servers:
  -
    url: 'http://localhost:8080'
    description: Generated server url
paths:
  /seasons:
    get:
      tags:
        - Seasons
      responses:
        '200':
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SeasonDto'
          description: OK
      operationId: getSeasons
      summary: Retrieve all seasons
    post:
      tags:
        - Seasons
      responses:
        '200':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SeasonDto'
          description: OK
      operationId: newSeason
      summary: End the current season and create the next one
  /players:
    get:
      tags:
        - Players
      responses:
        '200':
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PlayerDto'
          description: OK
      operationId: getAllPlayers
      summary: Retrieve all players
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlayerCreationDto'
        required: true
      tags:
        - Players
      responses:
        '400':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PlayerDto'
          description: Invalid Player
      operationId: addPlayer
      summary: Sign up a new player
  /games:
    get:
      tags:
        - Games
      responses:
        '200':
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GameDto'
          description: OK
      operationId: getGames
      summary: Retrieve all games
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameCreationDto'
        required: true
      tags:
        - Games
      responses:
        '200':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/GameDto'
          description: OK
      operationId: addGame
      summary: Report a new game
  /seasons/current:
    get:
      tags:
        - Seasons
      responses:
        '200':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SeasonDto'
          description: OK
      operationId: currentSeason
      summary: Retrieve current version
  '/players/{name}':
    get:
      tags:
        - Players
      parameters:
        -
          name: name
          description: Name of the player you are looking for
          schema:
            type: string
          in: path
          required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerDto'
          description: OK
        '404':
          description: Player not found
      operationId: getPlayer
      summary: Retrieve given player
  '/players/{name}/games':
    get:
      tags:
        - Players
        - Games
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GameDto'
          description: OK
      operationId: getAllGamesForPlayer
      summary: Retrieve all games a given player played
    parameters:
      -
        name: name
        description: Name of the player
        schema:
          type: string
        in: path
        required: true
components:
  schemas:
    SeasonDto:
      description: 'Represents a season. When a new season starts, all players rankings go back to 1200.'
      type: object
      properties:
        number:
          format: int64
          description: Number of the season. This can be used as an ID.
          type: integer
        start:
          description: First day of the season
          type: string
        end:
          description: Last day of the season. This can be empty for the current season.
          type: string
        rankings:
          description: List of all players rankings for the season.
          type: array
          items:
            $ref: '#/components/schemas/SeasonRankingDto'
      example:
        number: 2
        start: '2024-04-03'
        end: '2025-03-21'
        rankings:
          -
            player: Tim
            elo: 185
          -
            player: Mathilde
            elo: 1402
    SeasonRankingDto:
      description: Represents a given player ranking for the season
      type: object
      properties:
        player:
          description: Player name
          type: string
        elo:
          format: int32
          description: Elo score for the player
          type: integer
      example:
        player: Jon
        elo: 1218
    PlayerCreationDto:
      description: Data needed to register a new Player
      required:
        - name
      type: object
      properties:
        name:
          description: New player's name
          type: string
      example:
        name: '4'
    PlayerDto:
      description: Details of a player
      type: object
      properties:
        name:
          description: The player's name
          type: string
        avatar:
          description: Name of the avatar image file
          type: string
        lastRanking:
          $ref: '#/components/schemas/PlayerRankingDto'
          description: The player ranking for the current season
        rankings:
          description: History of all rankings per season
          type: array
          items:
            $ref: '#/components/schemas/PlayerRankingDto'
        nbGames:
          format: int32
          description: Total number of games played.
          type: integer
        nbWins:
          format: int32
          description: Total number of wins
          type: integer
        winRate:
          format: int32
          description: Global win rate
          type: integer
        nbWinsSeason:
          description: Number of wins in the current season
          type: integer
        nbGamesSeason:
          description: Number of games played in the current season.
          type: integer
        winRateSeason:
          description: Winrate in the current season
          type: integer
      example:
        name: Jon
        avatar: coucou.png
        lastRanking:
          season: 3
          elo: 1184
        rankings:
          -
            season: 1
            elo: 1223
          -
            season: 2
            elo: 1302
        nbGames: 70
        nbWins: 35
        winRate: 50
        nbWinsSeason: 10
        nbGamesSeason: 15
        winRateSeason: 66
    PlayerRankingDto:
      description: Represents a given season rating for the player
      type: object
      properties:
        season:
          format: int64
          description: Season number
          type: integer
        elo:
          format: int32
          description: Elo score for that season
          type: integer
      example:
        season: 4
        elo: 1250
    GameCreationDto:
      description: Data needed to register a new game
      required:
        - losers
        - winners
      type: object
      properties:
        winners:
          description: >-
            List of names of players on the winning side. All players have to be registered into
            the app before appearing on a game.
          uniqueItems: true
          type: array
          items:
            type: string
        losers:
          description: >-
            List of names of players on the losing side. All players have to be registered into
            the app before appearing on a game.
          uniqueItems: true
          type: array
          items:
            type: string
        losersScore:
          format: int32
          description: >-
            How many points (between 0 and 12 included) the losers scored before dropping the
            game.
          maximum: 12
          type: integer
      example:
        winners:
          - Olivier
          - Jon
        losers:
          - Tim
          - Garrett
        losersScore: 1
    GameDto:
      description: Represents a game.
      type: object
      properties:
        rank:
          format: int64
          description: 'Chronological rank of the game, since the beginning of times...'
          type: integer
        winners:
          description: >-
            List of names of players on the winning side. All players have to be registered into
            the app before appearing on a game.
          type: array
          items:
            type: string
        losers:
          description: >-
            List of names of players on the losing side. All players have to be registered into
            the app before appearing on a game.
          type: array
          items:
            type: string
        losersScore:
          format: int32
          description: >-
            How many points (between 0 and 12 included) the losers scored before dropping the
            game.
          type: integer
        eloSwitch:
          format: int32
          description: >-
            Impact the game had on every player ranking. This value represents how many points the
            winners earned, and therefore how many points got deduced from each loser's rankings.
          type: integer
        season:
          format: int64
          description: >-
            Identifier of the season this game belongs to. Seasons endpoints may be queried with
            this ID.
          type: integer
        gameDay:
          description: Date on which the game happened
          type: string
      example:
        rank: 11
        winners:
          - Olivier
          - Mylene
        losers:
          - Leny
          - Jon
        losersScore: 6
        eloSwitch: 19
        season: 2
        gameDay: '2024-04-25'